+++ agents-cli/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @inkeep/agents-cli
 
+## 0.64.4
+
+### Patch Changes
+
+- Updated dependencies [707333c]
+  - @inkeep/agents-manage-ui@0.64.4
+  - @inkeep/agents-core@0.64.4
+  - @inkeep/agents-sdk@0.64.4
+
+## 0.64.3
+
+### Patch Changes
+
+- Updated dependencies [d10ccf6]
+- Updated dependencies [63b1062]
+- Updated dependencies [7aa1fac]
+- Updated dependencies [7aa1fac]
+- Updated dependencies [4ace590]
+- Updated dependencies [ee1d23a]
+  - @inkeep/agents-manage-ui@0.64.3
+  - @inkeep/agents-core@0.64.3
+  - @inkeep/agents-sdk@0.64.3
+
+## 0.64.2
+
+### Patch Changes
+
+- Updated dependencies [f0081f7]
+- Updated dependencies [f099221]
+- Updated dependencies [a41b6c7]
+- Updated dependencies [8c16740]
+- Updated dependencies [b55675b]
+  - @inkeep/agents-manage-ui@0.64.2
+  - @inkeep/agents-core@0.64.2
+  - @inkeep/agents-sdk@0.64.2
+
 ## 0.64.1
 
 ### Patch Changes